agents/flashcard-gen/requirements.txt
streamlit
pypdf
langchain
langchain-openai
python-dotenv
